Check before dir before rm files

+++ update-index.sh
@@ -7,11 +7,8 @@ BLUE='\033[0;34m'
 PURPLE='\033[0;35m'
 NC='\033[0m' # No Color
 
-echo -e "\n${RED}Removing generated old files\n* rm -r public/\n* > content/projects/list-ids\n* > content/projects/list-names"
 # Remove anything old and empty lists
-cd public/
-rm -r `ls | grep -v ".git"`
-cd ..
+[ -d public ] && echo -e "\n${RED}Removing generated old files\n* rm -r public/\n* > content/projects/list-ids\n* > content/projects/list-names" && cd public/ && rm -r `ls | grep -v ".git"` && cd ..
 > content/projects/list-ids
 > content/projects/list-names
